requirements/source.txt
# The requirements checked out as git submodules in vendor/src
# Must be kept up-to-date at all times.

# NO EDITABLE/VERSION CONTROL REQUIREMENTS ALLOWED!

python-dateutil==2.4.2

sqlparse==0.1.18

cssselect==0.9.1

Werkzeug==0.11.2

django-devserver==0.8.0

amqp==1.4.7

basket-client==0.3.11

billiard==3.3.0.21

bleach==1.4.2

bitly_api==0.3

celery==3.1.19

elasticsearch-dsl==0.0.9

elasticsearch==1.9.0

hash_ring==1.3.1

html5lib==0.999999

jingo==0.7.1

kombu==3.0.29

oauthlib==1.0.3

python-magic==0.4.10

python-memcached==1.54

python-openid==2.2.5

pytidylib==0.2.4

# FIXME: not the newest version given errormill constraints
raven==3.3.12

requests==2.8.1

requests-mock==0.7.0

requests-oauthlib==0.5.0

puente==0.3

urllib3==1.12

URLObject==2.4.0

pyquery==1.2.9

Django==1.7.11

commonware==0.4.3

django-allauth==0.24.1

django-appconf==1.0.1

django-celery==3.1.17

django-constance==1.1.1

django-extensions==1.5.9

django-honeypot==0.5.0

django-memcached-hashring==0.1.2

django-nose==1.4.2

djangorestframework==3.3.1

django-smuggler==0.6.1

django-soapbox==1.2

django-statici18n==1.1.5

django-taggit==0.17.3

django-tidings==1.1

django-timezones==0.2

django-waffle==0.11

django-cacheback==0.9.1

django-debug-toolbar==1.4

django-mozilla-product-details==0.7.1

django-ratelimit==0.6.0

django-flat-theme==1.1.1

django-sundial==1.0.1

django-pipeline==1.5.4

django-celery-transactions==0.3.6
